+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/util/ClosureHelper.java
@@ -68,7 +68,11 @@ public class ClosureHelper {
         return invokeClosureWithDelegate(closureType, delegate, delegate);
     }
 
-    public static <T> T invokeClosureWithDelegate(Closure<T> closure,  Object delegate, Object... arguments) {
+    public static <T> T invokeClosureWithDelegateAsArgument(Closure<T> closure, Object delegate) {
+        return invokeClosureWithDelegate(closure, delegate, delegate);
+    }
+    public static <T> T invokeClosureWithDelegate(Closure<T> closure, Object delegate, Object... arguments) {
+        if (closure == null) return null;
         if (delegate != null) {
             closure.setResolveStrategy(Closure.DELEGATE_FIRST);
             closure.setDelegate(delegate);